+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/OrganizationRepository.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
  */
 
 @Repository
-public interface OrganizationRepository extends JpaRepository<Organization,Long> {
+public interface OrganizationRepository extends JpaRepository<Organization,Integer> {
 }
